hackzurich22-backend/requirements.txt
flask>=2.2.2
websockets>=10.3
asyncio>=3.4.3
websockets-routes>=0.2
snowflake-id